backend/requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
boto3==1.28.78
botocore==1.31.78
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
ecdsa==0.18.0
fastapi==0.104.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
hiredis==2.2.3
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==9.1.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.25.0
pyarrow==14.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-jose==3.3.0
python-ulid==1.1.0
pytz==2023.3.post1
redis==4.6.0
redis-om==0.2.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.28.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.9
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.24.0
validators==0.22.0
zipp==3.17.0
